add city

+++ app/src/main/java/com/kohuyn/weatherapp/data/model/currentweather/Sys.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.Expose
 data class Sys(
     @Expose val country: String,
     @Expose val id: Int,
-    @Expose val sunrise: Int,
-    @Expose val sunset: Int,
+    @Expose val sunrise: Long,
+    @Expose val sunset: Long,
     @Expose val type: Int
 )
